cx marcian art for errant opaque white pixels
corrected errant white pixels to artwork for marcia. This artwork is already 32 pix x 64 pixels.

some images have color palette anamolies that will not effect the game. They include transparent colors that are difficult to find within the composition. It was not originally intended that the colored pixels would have alpha. Only the background was to be transparent.